CU-868k9j8az :: Add Room Grouping functionality to ZoneEditor

+++ Assets/Scripts/Zones/Editor/ZoneEditor/ZoneGraphView.cs
@@ -10,14 +10,17 @@ namespace Frankie.ZoneManagement.Editor
     {
         // State
         private Zone zone;
-        private VisualElement canvasContent;
+        private readonly VisualElement canvasContent;
+        private readonly VisualElement groupLayer;
         private readonly ZoneEdgesLayer edgesLayer;
         private readonly VisualElement nodeLayer;
         private readonly StandardCanvasPanManipulator panManipulator;
         private readonly StandardCanvasZoomManipulator zoomManipulator;
         private readonly Dictionary<string, ZoneNodeView> nodeViewLookup = new();
+        private readonly Dictionary<ZoneNodeGroup, ZoneNodeGroupView> groupViewLookup = new();
         private ZoneNode linkingParentNode;
-        
+        private bool isPlacingGroup;
+
         public ZoneGraphView()
         {
             canvasContent = MakeCanvas();
@@ -26,12 +29,16 @@ namespace Frankie.ZoneManagement.Editor
             var backgroundLayer = new ZoneBackgroundLayer();
             canvasContent.Add(backgroundLayer);
 
+            groupLayer = MakeGroupLayer();
+            canvasContent.Add(groupLayer);
+
             edgesLayer = new ZoneEdgesLayer();
             canvasContent.Add(edgesLayer);
 
             nodeLayer = MakeNodesLayer();
             canvasContent.Add(nodeLayer);
 
+            RegisterCallback<MouseDownEvent>(OnPossibleGroupPlacementClick);
             panManipulator = new StandardCanvasPanManipulator(canvasContent);
             this.AddManipulator(panManipulator);
             zoomManipulator = new StandardCanvasZoomManipulator(canvasContent);
@@ -50,11 +57,14 @@ namespace Frankie.ZoneManagement.Editor
         {
             zone = newZone;
             linkingParentNode = null;
+            isPlacingGroup = false;
             RebuildNodes();
+            RebuildGroups();
         }
         
         public void RequestNodeIDChange(ZoneNode zoneNode, string newID)
         {
+            if (zone == null) { return; }
             if (string.IsNullOrWhiteSpace(newID)) { return; }
             if (zone.GetNodeFromID(newID) != null) { return; } // ID already taken
 
@@ -68,17 +78,48 @@ namespace Frankie.ZoneManagement.Editor
         
         public void RequestCreateChild(ZoneNode parentNode)
         {
+            if (zone == null) { return; }
             zone.CreateChildNode(parentNode);
             RebuildNodes();
         }
 
         public void RequestDelete(ZoneNode nodeToDelete)
         {
+            if (zone == null) { return; }
             zone.DeleteNode(nodeToDelete);
             RebuildNodes();
+            RefreshGroupViews();
         }
+
+        public bool TryGetZoneNodeView(string nodeID, out ZoneNodeView nodeView) => nodeViewLookup.TryGetValue(nodeID, out nodeView);
         
         public void NotifyNodeMoved() => RefreshEdges();
+
+        public void UpdateGroupsForNode(ZoneNode movedNode, bool forceRecalculation = false)
+        {
+            if (zone == null) { return; }
+            zone.UpdateGroupsForNodeMove(movedNode);
+            RefreshGroupViews(forceRecalculation);
+        }
+
+        public void BeginPlacingGroup()
+        {
+            if (zone == null) { return; }
+            isPlacingGroup = true;
+        }
+
+        public void RequestDeleteGroup(ZoneNodeGroup group)
+        {
+            if (zone == null) { return; }
+            zone.DeleteGroup(group);
+            RebuildGroups();
+        }
+
+        public void SetGroupRect(ZoneNodeGroup group, Rect rect)
+        {
+            if (zone == null) { return; }
+            zone.SetGroupRect(group, rect);
+        }
 
         public void BeginLinking(ZoneNode parentNode)
         {
@@ -94,6 +135,7 @@ namespace Frankie.ZoneManagement.Editor
 
         public void CompleteLinking(ZoneNode targetNode)
         {
+            if (zone == null) { return; }
             if (linkingParentNode == null) { return; }
             zone.ToggleRelation(linkingParentNode, targetNode);
             linkingParentNode = null;
@@ -152,6 +194,45 @@ namespace Frankie.ZoneManagement.Editor
                 nodeView.RefreshLinkButton();
             }
         }
+
+        private void RebuildGroups()
+        {
+            groupLayer.Clear();
+            groupViewLookup.Clear();
+            
+            if (zone == null) { return; }
+
+            foreach (ZoneNodeGroup zoneNodeGroup in zone.GetAllGroups())
+            {
+                var zoneNodeGroupView = new ZoneNodeGroupView(zoneNodeGroup, this);
+                groupViewLookup[zoneNodeGroup] = zoneNodeGroupView;
+                groupLayer.Add(zoneNodeGroupView);
+            }
+        }
+
+        private void RefreshGroupViews(bool forceRecalculation = false)
+        {
+            foreach (ZoneNodeGroupView zoneGroupView in groupViewLookup.Values.Where(zoneGroupView => zoneGroupView?.zoneNodeGroup != null))
+            {
+                if (forceRecalculation) { zoneGroupView.zoneNodeGroup.RecomputeGroupRect(); }
+                zoneGroupView.ApplyRectFromData();
+            }
+        }
+
+        private void OnPossibleGroupPlacementClick(MouseDownEvent mouseDownEvent)
+        {
+            if (zone == null) { return; }
+            if (!isPlacingGroup) { return; }
+            if (mouseDownEvent.button != (int)MouseButton.LeftMouse) { return; }
+
+            isPlacingGroup = false;
+            
+            Vector2 contentPosition = canvasContent.WorldToLocal(mouseDownEvent.mousePosition);
+
+            zone.CreateZoneNodeGroup(contentPosition);
+            RebuildGroups();
+            mouseDownEvent.StopImmediatePropagation();
+        }
         #endregion
         
         #region StaticUIBuilders
@@ -160,6 +241,20 @@ namespace Frankie.ZoneManagement.Editor
             return new VisualElement
             {
                 name = "canvas-content",
+                style =
+                {
+                    position = Position.Absolute,
+                    left = 0,
+                    top = 0
+                }
+            };
+        }
+
+        private static VisualElement MakeGroupLayer()
+        {
+            return new VisualElement
+            {
+                name = "zone-group-layer",
                 style =
                 {
                     position = Position.Absolute,
